risugen: Correct generated code for register alignment and flag clearing
The correct logical operation to use for clearing bits in a register
is an AND, not an XOR. Correct the code we emit to align registers
to use the right logic op (and the right immediate value). The code
emitted to clear the NZCV flags had a similar problem but the fix
here is even simpler: just write the zero register to NZCV rather
than trying to do a read-mask-write.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>